Consolidate CollectionsHelper and CurationConcerns::CollectionsHelper
