--- app/helpers/collections_helper.rb
@@ -1,4 +1,0 @@
-# View Helpers for Hydra Collections functionality
-module CollectionsHelper
-  include CurationConcerns::CollectionsHelperBehavior
-end
